Better shellcheck tooling
Adds a script for developers that builds a given xdg-util using the Makefile, shellchecks it and calculates the lines in the source file.

This should allow easier shellchecking of the final output and avoid getting used to false positives.

This also disables SC3043 (`local <name>` statements without an declaration) which causes more noise than necessary.