force explicit session id in xss-lock
this is the only clean, or as clean as it gets way to detect that xss-lock
is running on the current session, this will force every user of xss-lock to
closely follow the example from /usr/share/doc/xss-lock/xss-lock.service and
explicitly run xss-lock on their system with the -s or --session flag so that
xdg-screensaver can detect the program is running on the current session.

in regards of reliability, if the $XDG_SESSION_ID variable is not present on
the current session that means it was malformed and issues ARE going to EXIST
regardless.