Made the .shellcheckrc file a bit more self-explaining
This is mostly to help out curious people who aren't familiar with shellcheck
and to improve readability for everyone else.

It also now communicates the intend of which rules are permanent
because of how the xdg-utils are built and which are there to
make the shellcheck output less cluttered while working on more
important issues.