xdg-open: support toolbx
Toolbx [1] is an utility that maps parts of your host system, such as
your regular home directory, into a containerized environment to
provide a seamless development experience; e.g. [2].

It is admittedly a bit of a confusing situation for something like
xdg-open, because you are running a container with parts of an
existing system mapped into it.  It's by design that some parts are
internal to the container (dev tools, etc.) and some parts are not
(homedirs, etc.) - there is some discussion at [3].  That said, since
this is very much designed for interactive/human use, when you have a
terminal running a toolbx development environment, your a priori
expectation is that something like an xdg-open call will open in your
host system browser.  This is essentially the same as the flatpak
situation, where the correct thing to do is use gdbus.

This change adds a check for the toolbox stamp file that tells us we
are running in a toolbox.  xdg-[open|email] are updated to use gdbus
for toolbox and flatpak, the function is renamed to make that clearer.

[1] https://containertoolbx.org/
[2] https://docs.fedoraproject.org/en-US/fedora-silverblue/toolbox/
[3] https://github.com/containers/toolbox/issues/291

Signed-off-by: Ian Wienand <iwienand@redhat.com>